Removed Old Kit-kat Code
Substituted old kit-kat code for modern lollipop session handling.
Should fix incorrect music player listed.
