+++ app/src/main/java/com/adam/aslfms/receiver/BootReceiver.java
@@ -6,38 +6,19 @@ import android.content.Intent;
 import android.os.Build;
 import android.util.Log;
 
-import com.adam.aslfms.service.ControllerReceiverCallback;
-import com.adam.aslfms.service.ControllerReceiverService;
-import com.adam.aslfms.service.NotificationBarService;
-import com.adam.aslfms.service.ScrobblingService;
+import com.adam.aslfms.util.Util;
 
 public class BootReceiver extends BroadcastReceiver {
 
-    private ControllerReceiverCallback controllerCallback = null;
     private static final String TAG = "BootReceiver";
-
     public static final String SYSTEM_ACTION = "android.intent.action.BOOT_COMPLETED";
 
     @Override
     public void onReceive(Context context, Intent intent) {
         if( Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            Log.d(TAG, "launching");
-            Intent i = new Intent(context, NotificationBarService.class);
+            Log.d(TAG, "launching_on_boot");
             if (intent != null || intent.getAction() == SYSTEM_ACTION) {
-                i.setAction(NotificationBarService.ACTION_NOTIFICATION_BAR_UPDATE);
-                i.putExtra("track", "");
-                i.putExtra("artist", "");
-                i.putExtra("album", "");
-                i.putExtra("app_name", "");
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                    context.startForegroundService(i);
-                    context.startForegroundService(new Intent(context, ScrobblingService.class));
-                    context.startForegroundService(new Intent(context, ControllerReceiverService.class));
-                } else {
-                    context.startService(i);
-                    context.startService(new Intent(context, ScrobblingService.class));
-                    context.startService(new Intent(context, ControllerReceiverService.class));
-                }
+                Util.runServices(context);
             }
         }
     }
