+++ app/src/main/java/com/adam/aslfms/service/ControllerReceiverService.java
@@ -3,45 +3,42 @@ package com.adam.aslfms.service;
 import android.annotation.TargetApi;
 import android.app.Service;
 import android.content.ComponentName;
-import android.content.ContentResolver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
-import android.media.AudioManager;
-import android.media.MediaMetadataEditor;
-import android.media.MediaMetadataRetriever;
-import android.media.RemoteControlClient;
-import android.media.RemoteController;
+import android.media.MediaMetadata;
+import android.media.session.MediaController;
+import android.media.session.MediaSessionManager;
+import android.media.session.PlaybackState;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
 import android.os.IBinder;
-import android.provider.Settings;
+import android.preference.PreferenceManager;
+import android.service.notification.NotificationListenerService;
 import android.support.annotation.RequiresApi;
-import android.support.v4.app.NotificationManagerCompat;
 import android.util.Log;
 
+import com.adam.aslfms.receiver.GenericControllerReceiver;
 import com.adam.aslfms.util.AppSettings;
 import com.adam.aslfms.util.NotificationCreator;
 import com.adam.aslfms.util.Util;
 
-import java.lang.ref.WeakReference;
-import java.util.Set;
+import java.util.List;
 
 @SuppressWarnings("deprecation")
 @TargetApi(Build.VERSION_CODES.LOLLIPOP)
 @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
-public class ControllerReceiverService extends android.service.notification.NotificationListenerService implements RemoteController.OnClientUpdateListener {
+public class ControllerReceiverService extends NotificationListenerService {
     
     private static final String TAG = "ControllerReceiverSrvc";
-    private static WeakReference<RemoteController> mRemoteController = new WeakReference<>(null);
-    private ControllerReceiverCallback controllerReceiverCallback;
-    private String track;
-    private String artist;
-    private String album;
-    private String albumArtist;
-    private Object durationObject;
+    private Handler handler = new Handler();
+    private String mPlayer = null;
+
+    private MediaSessionManager mediaSessionManager;
+    private ComponentName componentName;
+    private MediaController controller;
+    private List<MediaController> controllers;
 
     @Override
     public IBinder onBind(Intent intent) {
@@ -52,35 +49,35 @@ public class ControllerReceiverService extends android.service.notification.Noti
     @SuppressWarnings("NewApi")
     public void onCreate() {
         super.onCreate();
+        Log.d(TAG,"created");
         AppSettings settings = new AppSettings(this);
         
-        registerControllerReceiverCallback();
-
+        //registerControllerReceiverCallback();
         Bundle extras = new Bundle();
-        extras.putString("track", track);
-        extras.putString("artist", artist);
-        extras.putString("album", album);
-        extras.putString("app_name", albumArtist);
+        extras.putString("track", "");
+        extras.putString("artist", "");
+        extras.putString("album", "");
+        extras.putString("app_name", "");
         this.startForeground(NotificationCreator.FOREGROUND_ID, NotificationCreator.prepareNotification(extras, this));
+        initMediaListener();
         if (!settings.isActiveAppEnabled(Util.checkPower(this))) {
-            this.stopForeground(true); // TODO: test if this conflicts/stops scrobbles
+            this.stopForeground(true);
         }
     }
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
+        Log.d(TAG,"started");
         AppSettings settings = new AppSettings(this);
-
-        registerControllerReceiverCallback();
-
         Bundle extras = new Bundle();
-        extras.putString("track", track);
-        extras.putString("artist", artist);
-        extras.putString("album", album);
-        extras.putString("app_name", albumArtist);
+        extras.putString("track", "");
+        extras.putString("artist", "");
+        extras.putString("album", "");
+        extras.putString("app_name", "");
         this.startForeground(NotificationCreator.FOREGROUND_ID, NotificationCreator.prepareNotification(extras, this));
+        initMediaListener();
         if (!settings.isActiveAppEnabled(Util.checkPower(this))) {
-            this.stopForeground(true); // TODO: test if this conflicts/stops scrobbles
+            this.stopForeground(true);
             return Service.START_NOT_STICKY;
         }
         return Service.START_STICKY;
@@ -89,108 +86,7 @@ public class ControllerReceiverService extends android.service.notification.Noti
     @Override
     public void onDestroy() {
         super.onDestroy();
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-            if (mRemoteController != null && mRemoteController.get() != null)
-                ((AudioManager) getSystemService(Context.AUDIO_SERVICE)).unregisterRemoteController(mRemoteController.get());
-        }
-    }
-
-    public static boolean isControllerReceiverServiceEnabled(Context context) {
-        Set<String> packageNames = NotificationManagerCompat.getEnabledListenerPackages(context);
-        return packageNames.contains(context.getPackageName());
-    }
-
-    private void disableControllerReceiverService() {
-        PackageManager pm = getPackageManager();
-        pm.setComponentEnabledSetting(new ComponentName(getApplicationContext(), ControllerReceiverService.class),
-                PackageManager.COMPONENT_ENABLED_STATE_DISABLED, PackageManager.DONT_KILL_APP);
-    }
-
-    /* KitKat stuff */
-
-    private boolean isRemoteControllerPlaying;
-    private boolean mHasBug = true;
-
-    @Override
-    public void onClientChange(boolean clearing) {
-        // isListeningAuthorized(ControllerReceiverService.this);
-    }
-
-    @Override
-    public void onClientPlaybackStateUpdate(int state) {
-        this.isRemoteControllerPlaying = state == RemoteControlClient.PLAYSTATE_PLAYING;
-    }
-
-
-    public static long getRemotePlayerPosition() {
-        return mRemoteController.get() != null ?
-                Math.min(3600000,mRemoteController.get().getEstimatedMediaPosition()) : -1L;
-    }
-
-    @Override
-    public void onClientPlaybackStateUpdate(int state, long stateChangeTimeMs, long currentPosMs, float speed) {
-        this.isRemoteControllerPlaying = state == RemoteControlClient.PLAYSTATE_PLAYING;
-        mHasBug = false;
-        if (currentPosMs > 3600000)
-            currentPosMs = -1L;
-        SharedPreferences current = getSharedPreferences("current_music", Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = current.edit();
-        editor.putLong("position", currentPosMs);
-
-        if (isRemoteControllerPlaying) {
-            long currentTime = System.currentTimeMillis();
-            editor.putLong("startTime", currentTime);
-        }
-        editor.putBoolean("playing", isRemoteControllerPlaying);
-        editor.apply();
-
-        Log.d(TAG, "PlaybackStateUpdate - position stored: " + currentPosMs);
-
-        long position = getRemotePlayerPosition();
-
-        if (durationObject instanceof Double) {
-            Log.d(TAG,"duration is Double");
-            if (artist != null && !artist.isEmpty())
-                controllerReceiverCallback.broadcast(this, artist, track, album, isRemoteControllerPlaying, (Double) durationObject, position, albumArtist, null);
-        } else if (durationObject instanceof Long) {
-            Log.d(TAG,"duration is Long");
-            if (artist != null && !artist.isEmpty())
-                controllerReceiverCallback.broadcast(this, artist, track, album, isRemoteControllerPlaying, (Long) durationObject, position, albumArtist, null);
-        } else if (durationObject instanceof Integer) {
-            Log.d(TAG, "duration is Integer");
-            if (artist != null && !artist.isEmpty())
-                controllerReceiverCallback.broadcast(this, artist, track, album, isRemoteControllerPlaying, (Integer) durationObject, position, albumArtist, null);
-        }
-    }
-
-    @Override
-    public void onClientTransportControlUpdate(int transportControlFlags) {
-        if (mHasBug && mRemoteController.get() != null) {
-            long position = mRemoteController.get().getEstimatedMediaPosition();
-            if (position > 3600000)
-                position = -1L;
-            SharedPreferences current = getSharedPreferences("current_music", Context.MODE_PRIVATE);
-            current.edit().putLong("position", position).apply();
-            if (isRemoteControllerPlaying) {
-                long currentTime = System.currentTimeMillis();
-                current.edit().putLong("startTime", currentTime).apply();
-            }
-            Log.d(TAG, "TransportControlUpdate - position stored: " + position);
-        }
-    }
-
-    @Override
-    public void onClientMetadataUpdate(RemoteController.MetadataEditor metadataEditor) {
-        // isRemoteControllerPlaying = true;
-
-        durationObject = metadataEditor.getObject(MediaMetadataRetriever.METADATA_KEY_DURATION, 1200); //allow it to pass if not present
-        artist = metadataEditor.getString(MediaMetadataRetriever.METADATA_KEY_ARTIST, "");
-        albumArtist = metadataEditor.getString(MediaMetadataRetriever.METADATA_KEY_ALBUMARTIST, "");
-        track = metadataEditor.getString(MediaMetadataRetriever.METADATA_KEY_TITLE, "");
-        album = metadataEditor.getString(MediaMetadataRetriever.METADATA_KEY_ALBUM, "");
-        Bitmap artwork = metadataEditor.getBitmap(MediaMetadataEditor.BITMAP_KEY_ARTWORK, null);
-
-        controllerReceiverCallback.saveArtwork(this, artwork, artist, track, album);
+        Log.d(TAG,"destroyed");
     }
 
     // BEGIN listener stuff
@@ -207,35 +103,124 @@ public class ControllerReceiverService extends android.service.notification.Noti
         requestRebind(new ComponentName(getApplicationContext(), ControllerReceiverService.class));
     }
 
-    public static boolean isListeningAuthorized(Context context) {
-        ContentResolver contentResolver = context.getContentResolver();
-        String enabledNotificationListeners = Settings.Secure.getString(contentResolver, "enabled_notification_listeners");
-        String packageName = context.getPackageName();
 
-        return !(enabledNotificationListeners == null || !enabledNotificationListeners.contains(packageName));
-    }
+    public void initMediaListener(){
+        componentName = new ComponentName(this, ControllerReceiverService.class);
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.LOLLIPOP) {
+            mediaSessionManager = (MediaSessionManager) getSystemService(Context.MEDIA_SESSION_SERVICE);
+            mediaSessionManager.addOnActiveSessionsChangedListener(sessionsChangedListener, componentName);
 
-    public static boolean isNotificationListenerServiceEnabled(Context context) {
-        Set<String> packageNames = NotificationManagerCompat.getEnabledListenerPackages(context);
-        return packageNames.contains(context.getPackageName());
-    }
-
-    public void registerControllerReceiverCallback(){
-        ControllerReceiverCallback controllerCallback = new ControllerReceiverCallback();
-        if (controllerCallback == null)
-            controllerCallback = new ControllerReceiverCallback();
-        if (ControllerReceiverService.isListeningAuthorized(this))
-            ControllerReceiverCallback.registerFallbackControllerCallback(this, controllerCallback);
-
-        if (NotificationManagerCompat.getEnabledListenerPackages (getApplicationContext()).contains(getApplicationContext().getPackageName())) {
-            mRemoteController = new WeakReference<>(new RemoteController(this, this));
-            mRemoteController.get().setArtworkConfiguration(3000, 3000);
-            if (!((AudioManager) getSystemService(Context.AUDIO_SERVICE)).registerRemoteController(mRemoteController.get())) {
-                throw new RuntimeException("Error while registering RemoteController!");
-            }
-            controllerReceiverCallback = new ControllerReceiverCallback();
+            this.controllers = mediaSessionManager.getActiveSessions(componentName);
         }
     }
 
-    // END listener stuff
+    MediaSessionManager.OnActiveSessionsChangedListener sessionsChangedListener = new MediaSessionManager.OnActiveSessionsChangedListener() {
+        public void onActiveSessionsChanged(List<MediaController> controllers) {
+            MediaController controller;
+            MediaController.Callback controllerCallback;
+            Log.d(TAG, "onActiveSessionsChanged");
+            if (controllers.size() > 0 && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                controller = controllers.get(0);
+                controllerCallback = new MediaController.Callback() {
+                    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+                    @Override
+                    public void onPlaybackStateChanged(PlaybackState state) {
+                        super.onPlaybackStateChanged(state);
+                        if (state == null)
+                            return;
+                        boolean isPlaying = state.getState() == PlaybackState.STATE_PLAYING;
+                        broadcastControllerState(controller, isPlaying);
+                    }
+
+                    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+                    @Override
+                    public void onMetadataChanged(MediaMetadata metadata) {
+                        super.onMetadataChanged(metadata);
+                        if (metadata == null)
+                            return;
+                        broadcastControllerState(controller, null);
+                    }
+                };
+                controller.registerCallback(controllerCallback);
+            }
+        }
+    };
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public void broadcastControllerState(MediaController mController, Boolean isPlaying) {
+        final MediaController mediaController = mController;
+        final Boolean playState = isPlaying;
+        handler.postDelayed(() -> {
+            mPlayer = mediaController.getPackageName();
+            MediaMetadata metadata = mediaController.getMetadata();
+            PlaybackState playbackState = mediaController.getPlaybackState();
+            if (metadata == null)
+                return;
+            String artist = null;
+            try {
+                artist = metadata.getString(MediaMetadata.METADATA_KEY_ARTIST);
+                if (artist == null)
+                    artist = metadata.getString(MediaMetadata.METADATA_KEY_ALBUM_ARTIST);
+            } catch (Exception ignored) {
+            }
+            String albumArtist = null;
+            try {
+                albumArtist = metadata.getString(MediaMetadata.METADATA_KEY_ALBUM_ARTIST);
+            } catch (Exception ignored){
+            }
+            String track = null;
+            try {
+                track = metadata.getString(MediaMetadata.METADATA_KEY_TITLE);
+            } catch (Exception ignored) {
+            }
+            String album = null;
+            try {
+                album = metadata.getString(MediaMetadata.METADATA_KEY_ALBUM);
+            } catch (Exception ignored) {
+            }
+            Bitmap artwork = null;
+            try {
+                artwork = metadata.getBitmap(MediaMetadata.METADATA_KEY_ALBUM_ART);
+                if (artwork == null)
+                    artwork = metadata.getBitmap(MediaMetadata.METADATA_KEY_ART);
+            } catch (Exception ignored) {
+            }
+
+            double duration;
+            try {
+                duration =(double) metadata.getLong(MediaMetadata.METADATA_KEY_DURATION);
+            } catch (RuntimeException ignored) {
+                duration = 0;
+            }
+            long position = duration == 0 || playbackState == null ? -1 : playbackState.getPosition();
+
+            if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("pref_filter_20min", true) && duration > 1200000)
+                return;
+            boolean tempPlaying = false;
+            if (playState == null)
+                tempPlaying = playbackState != null && playbackState.getState() == PlaybackState.STATE_PLAYING;
+
+            String player = mediaController.getPackageName();
+            if ("com.aimp.player".equals(player)) // Aimp is awful
+                position = -1;
+            broadcast(artist, track, album, tempPlaying, duration, position, albumArtist, player);
+        }, 50);
+        Log.d(TAG, "broadcast sent: controller state");
+    }
+
+    public void broadcast(String artist, String track, String album, boolean playing, double duration, long position, String albumArtist, String player) {
+        Intent localIntent = new Intent(GenericControllerReceiver.ACTION_INTENT);
+        localIntent.setComponent(new ComponentName(this.getPackageName(),"com.adam.aslfms.receiver.GenericControllerReceiver"));
+        localIntent.putExtra("artist", artist);
+        localIntent.putExtra("track", track);
+        localIntent.putExtra("album", album);
+        localIntent.putExtra("albumArtist", albumArtist);
+        localIntent.putExtra("playing", playing);
+        localIntent.putExtra("duration", duration);
+        localIntent.putExtra("player", mPlayer);
+        if (position != -1)
+            localIntent.putExtra("position", position);
+        this.sendBroadcast(localIntent);
+        Log.d(TAG, "broadcast sent: double duration");
+    }
 }
