+++ app/src/main/java/com/adam/aslfms/service/NotificationBarService.java
@@ -41,7 +41,7 @@ public class NotificationBarService extends Service {
 
         this.startForeground(NotificationCreator.FOREGROUND_ID, NotificationCreator.prepareNotification(extras, mCtx));
         if (!settings.isActiveAppEnabled(Util.checkPower(mCtx))) {
-            this.stopForeground(true); // TODO: test if this conflicts/stops scrobbles
+            this.stopForeground(true);
         }
     }
 
@@ -55,7 +55,7 @@ public class NotificationBarService extends Service {
         handleCommand(i, startId);
         this.startForeground(NotificationCreator.FOREGROUND_ID, NotificationCreator.prepareNotification(extras, mCtx));
         if (!settings.isActiveAppEnabled(Util.checkPower(mCtx))) {
-            this.stopForeground(true); // TODO: test if this conflicts/stops scrobbles
+            this.stopForeground(true);
             return Service.START_NOT_STICKY;
         }
         return Service.START_STICKY;
