+++ app/src/main/java/com/adam/aslfms/service/ScrobblingService.java
@@ -97,7 +97,7 @@ public class ScrobblingService extends Service {
         }
         this.startForeground(NotificationCreator.FOREGROUND_ID, NotificationCreator.prepareNotification(extras, mCtx));
         if (!settings.isActiveAppEnabled(Util.checkPower(mCtx))) {
-            this.stopForeground(true); // TODO: test if this conflicts/stops scrobbles
+            this.stopForeground(true);
         }
     }
 
@@ -123,7 +123,7 @@ public class ScrobblingService extends Service {
         }
         this.startForeground(NotificationCreator.FOREGROUND_ID, NotificationCreator.prepareNotification(extras, mCtx));
         if (!settings.isActiveAppEnabled(Util.checkPower(mCtx))) {
-            this.stopForeground(true); // TODO: test if this conflicts/stops scrobbles
+            this.stopForeground(true);
             return Service.START_NOT_STICKY;
         }
         return Service.START_STICKY;
